refactor: Added cache class for sdk use.
